+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 1.0.6 (in development)
 
+- Module 8FACE
+    - Prevent storing presets from modules that use parameter mapping because of engine deadlock
 - Module MIDI-CAT
     - New module, mapping module similar to VCV's MIDI-MAP with midi feedback and note mapping
 - Module STRIP